feat: add unoptimized purchase summary card
